@@ -6,9 +6,11 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.sansoft.butterflies.R
-import kotlinx.android.synthetic.main.endangered_fragment.*
+import com.sansoft.butterflies.databinding.EndangeredFragmentBinding
 
 class EndangeredFragment : Fragment() {
+    private var _binding: EndangeredFragmentBinding? = null
+    private val binding get() = _binding!!
 
      val layoutResource: Int
         get() = R.layout.endangered_fragment
@@ -16,14 +18,16 @@ class EndangeredFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(layoutResource, container, false)
+    ): View {
+        _binding = EndangeredFragmentBinding.inflate(inflater, container, false)
+
+        return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
 
-        var stream = context?.assets?.open("species.pdf")
-        pdfView_endangered.fromStream(stream!!)
+        val stream = context?.assets?.open("species.pdf")
+        binding.pdfViewEndangered.fromStream(stream!!)
             .password(null) // if password protected, then write password
             .defaultPage(0) // set the default page to open
             .enableSwipe(true)
