@@ -4,18 +4,18 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.appcompat.app.AppCompatActivity
 import androidx.constraintlayout.widget.ConstraintLayout
 import com.sansoft.butterflies.R
+import com.sansoft.butterflies.databinding.MainMenuFragmentBinding
+import gr.jkapsouras.butterfliesofgreece.MainActivity
 import gr.jkapsouras.butterfliesofgreece.base.BaseFragment
 import gr.jkapsouras.butterfliesofgreece.base.UiComponent
-import kotlinx.android.synthetic.main.activity_main.*
-import kotlinx.android.synthetic.main.main_menu_fragment.*
 import org.koin.android.ext.android.inject
 
 
 class MainMenuFragment : BaseFragment<MenuPresenter>() {
-
+    private var _binding: MainMenuFragmentBinding? = null
+    private val binding get() = _binding!!
     override val layoutResource: Int
         get() = R.layout.main_menu_fragment
     override val presenter: MenuPresenter by inject()
@@ -28,44 +28,42 @@ class MainMenuFragment : BaseFragment<MenuPresenter>() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.main_menu_fragment, container, false)
+    ): View {
+        _binding = MainMenuFragmentBinding.inflate(inflater, container, false)
+
+        return binding.root
     }
 
     override fun onStart() {
         super.onStart()
-        (requireActivity() as AppCompatActivity).toolbar.visibility = View.GONE
+        (requireActivity() as MainActivity).binding.toolbar.visibility = View.GONE
     }
 
     override fun onResume() {
         super.onResume()
-        (requireActivity() as AppCompatActivity).toolbar.visibility = View.GONE
+        (requireActivity() as MainActivity).binding.toolbar.visibility = View.GONE
     }
 
     override fun onStop() {
         super.onStop()
-        (requireActivity() as AppCompatActivity).toolbar.visibility = View.VISIBLE
+        (requireActivity() as MainActivity).binding.toolbar.visibility = View.VISIBLE
     }
 
     override fun onPause() {
         super.onPause()
-        (requireActivity() as AppCompatActivity).toolbar.visibility = View.VISIBLE
-    }
-
-    override fun initView(view: View): View {
-        return view
+        (requireActivity() as MainActivity).binding.toolbar.visibility = View.VISIBLE
     }
 
     override fun initializeComponents(constraintLayout: ConstraintLayout): List<UiComponent> {
 
         menuComponent = MenuComponent(
-            btn_field,
-            btn_contribute,
-            btn_about,
-            btn_introduction,
-            btn_endangered,
-            btn_legal,
-            btn_recognition
+            binding.btnField,
+            binding.btnContribute,
+            binding.btnAbout,
+            binding.btnIntroduction,
+            binding.btnEndangered,
+            binding.btnLegal,
+            binding.btnRecognition
         )
 
         return listOf(menuComponent!!)
