@@ -1,18 +1,22 @@
 package gr.jkapsouras.butterfliesofgreece.fragments.photos
 
+import android.os.Bundle
+import android.view.LayoutInflater
 import android.view.View
+import android.view.ViewGroup
 import androidx.constraintlayout.widget.ConstraintLayout
 import com.sansoft.butterflies.R
+import com.sansoft.butterflies.databinding.PhotosFragmentBinding
 import gr.jkapsouras.butterfliesofgreece.base.BaseFragment
 import gr.jkapsouras.butterfliesofgreece.base.UiComponent
 import gr.jkapsouras.butterfliesofgreece.fragments.photos.components.PhotosCollectionComponent
 import gr.jkapsouras.butterfliesofgreece.fragments.photos.components.PhotosHeaderComponent
 import gr.jkapsouras.butterfliesofgreece.fragments.photos.components.PhotosTableComponent
-import kotlinx.android.synthetic.main.photos_fragment.*
 import org.koin.android.ext.android.inject
 
 class PhotosFragment : BaseFragment<PhotosPresenter>() {
-
+    private var _binding: PhotosFragmentBinding? = null
+    private val binding get() = _binding!!
     override val presenter: PhotosPresenter by inject()
     private var photosTableComponent: PhotosTableComponent? = null
     private var photosHeaderComponent: PhotosHeaderComponent? = null
@@ -21,14 +25,20 @@ class PhotosFragment : BaseFragment<PhotosPresenter>() {
     override val layoutResource: Int
         get() = R.layout.photos_fragment
 
-    override fun initView(view: View): View {
-        return view
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = PhotosFragmentBinding.inflate(inflater, container, false)
+
+        return binding.root
     }
 
     override fun initializeComponents(constraintLayout: ConstraintLayout): List<UiComponent> {
-        photosTableComponent = PhotosTableComponent(view_photos_table_photos)
-        photosCollectionComponent = PhotosCollectionComponent(view_photos_collection_photos)
-        photosHeaderComponent = PhotosHeaderComponent(view_photos_header)
+        photosTableComponent = PhotosTableComponent(binding.viewPhotosTablePhotos)
+        photosCollectionComponent = PhotosCollectionComponent(binding.viewPhotosCollectionPhotos)
+        photosHeaderComponent = PhotosHeaderComponent(binding.viewPhotosHeader)
         return listOf(photosTableComponent!!, photosCollectionComponent!!, photosHeaderComponent!!)
     }
 
