@@ -3,6 +3,7 @@ package gr.jkapsouras.butterfliesofgreece.fragments.previewer.uiEvents
 import gr.jkapsouras.butterfliesofgreece.base.UiEvent
 
 sealed class PdfPreviewEvents: UiEvent() {
-    object SharePdf : PdfPreviewEvents()
-    object CreatePdf : PdfPreviewEvents()
+    data object SharePdf : PdfPreviewEvents()
+    data object CreatePdf : PdfPreviewEvents()
+    data object DeleteData : PdfPreviewEvents()
 }
