@@ -4,12 +4,12 @@ import android.content.ContentValues
 import android.content.Context
 import android.util.Log
 import android.util.TypedValue
-import android.view.View
 import androidx.constraintlayout.widget.Constraints
 import androidx.core.content.ContextCompat
 import androidx.core.content.res.ResourcesCompat
 import androidx.recyclerview.widget.RecyclerView
 import com.sansoft.butterflies.R
+import com.sansoft.butterflies.databinding.RowPhotosCollectionBinding
 import gr.jkapsouras.butterfliesofgreece.base.UiEvent
 import gr.jkapsouras.butterfliesofgreece.dto.ButterflyPhoto
 import gr.jkapsouras.butterfliesofgreece.dto.Family
@@ -19,9 +19,8 @@ import gr.jkapsouras.butterfliesofgreece.fragments.photos.uiEvents.PhotosEvents
 import gr.jkapsouras.butterfliesofgreece.fragments.species.uiEvents.SpeciesEvents
 import gr.jkapsouras.butterfliesofgreece.views.photosTableView.adapter.ShowingStep
 import io.reactivex.rxjava3.subjects.PublishSubject
-import kotlinx.android.synthetic.main.row_photos_collection.view.*
 
-class PhotosCollectionViewCell(itemView: View, private val emitter: PublishSubject<UiEvent>) : RecyclerView.ViewHolder(itemView){
+class PhotosCollectionViewCell(val binding: RowPhotosCollectionBinding, private val emitter: PublishSubject<UiEvent>) : RecyclerView.ViewHolder(binding.root){
 
     var familyId:Int = -1
     var specieId:Int = -1
@@ -44,7 +43,7 @@ class PhotosCollectionViewCell(itemView: View, private val emitter: PublishSubje
             Log.d(Constraints.TAG, "field clicked")
         }
 
-        itemView.image_add_collection.setOnClickListener {
+        binding.imageAddCollection.setOnClickListener {
             when (showingStep) {
                 ShowingStep.Families ->
                     emitter.onNext(FamilyEvents.AddPhotosForPrintClicked(familyId))
@@ -64,7 +63,7 @@ class PhotosCollectionViewCell(itemView: View, private val emitter: PublishSubje
 
     fun update(family: Family, showingStep: ShowingStep){
         this.showingStep = showingStep
-        itemView.label_photo_collection_name.text = family.name
+        binding.labelPhotoCollectionName.text = family.name
         familyId = family.id
         try {
             val iden = itemView.resources.getIdentifier(
@@ -75,12 +74,12 @@ class PhotosCollectionViewCell(itemView: View, private val emitter: PublishSubje
             iden.let { tmpiden ->
                 val drawable = ResourcesCompat.getDrawable(itemView.resources, tmpiden, null)
                 drawable.let{
-                    itemView.iv_image_row_collection.setImageResource(tmpiden)
+                    binding.ivImageRowCollection.setImageResource(tmpiden)
                 }
             }
         }
         catch (e: Exception){
-            itemView.iv_image_row_collection.setImageDrawable(
+            binding.ivImageRowCollection.setImageDrawable(
                 ContextCompat.getDrawable(
                     itemView.context,
                     R.drawable.thumb_default
@@ -88,9 +87,9 @@ class PhotosCollectionViewCell(itemView: View, private val emitter: PublishSubje
             )
         }
 
-        itemView.cardView_image_row_collection.post {
-            val w = itemView.cardView_image_row_collection.width
-            itemView.cardView_image_row_collection.apply {
+        binding.cardViewImageRowCollection.post {
+            val w = binding.cardViewImageRowCollection.width
+            binding.cardViewImageRowCollection.apply {
                 radius = (w/2).toDp(itemView.context).toFloat()
             }
         }
@@ -98,7 +97,7 @@ class PhotosCollectionViewCell(itemView: View, private val emitter: PublishSubje
 
     fun update(specie: Specie, showingStep: ShowingStep){
         this.showingStep = showingStep
-        itemView.label_photo_collection_name.text = specie.name
+        binding.labelPhotoCollectionName.text = specie.name
         specieId = specie.id
         try {
             val iden = itemView.resources.getIdentifier(
@@ -109,12 +108,12 @@ class PhotosCollectionViewCell(itemView: View, private val emitter: PublishSubje
             iden.let { tmpiden ->
                 val drawable = ResourcesCompat.getDrawable(itemView.resources, tmpiden, null)
                 drawable.let{
-                    itemView.iv_image_row_collection.setImageResource(tmpiden)
+                    binding.ivImageRowCollection.setImageResource(tmpiden)
                 }
             }
         }
         catch (e: Exception){
-            itemView.iv_image_row_collection.setImageDrawable(
+            binding.ivImageRowCollection.setImageDrawable(
                 ContextCompat.getDrawable(
                     itemView.context,
                     R.drawable.thumb_default
@@ -122,9 +121,9 @@ class PhotosCollectionViewCell(itemView: View, private val emitter: PublishSubje
             )
         }
 
-        itemView.cardView_image_row_collection.post {
-            val w = itemView.cardView_image_row_collection.width
-            itemView.cardView_image_row_collection.apply {
+        binding.cardViewImageRowCollection.post {
+            val w = binding.cardViewImageRowCollection.width
+            binding.cardViewImageRowCollection.apply {
                 radius = (w/2).toDp(itemView.context).toFloat()
             }
         }
@@ -132,7 +131,7 @@ class PhotosCollectionViewCell(itemView: View, private val emitter: PublishSubje
 
     fun update(photo: ButterflyPhoto, showingStep:ShowingStep){
         this.showingStep = showingStep
-        itemView.label_photo_collection_name.text = photo.author
+        binding.labelPhotoCollectionName.text = photo.author
         photoId = photo.id
         try {
             val iden = itemView.resources.getIdentifier(
@@ -143,12 +142,12 @@ class PhotosCollectionViewCell(itemView: View, private val emitter: PublishSubje
             iden.let { tmpiden ->
                 val drawable = ResourcesCompat.getDrawable(itemView.resources, tmpiden, null)
                 drawable.let{
-                    itemView.iv_image_row_collection.setImageResource(tmpiden)
+                    binding.ivImageRowCollection.setImageResource(tmpiden)
                 }
             }
         }
         catch (e: Exception){
-            itemView.iv_image_row_collection.setImageDrawable(
+            binding.ivImageRowCollection.setImageDrawable(
                 ContextCompat.getDrawable(
                     itemView.context,
                     R.drawable.thumb_default
@@ -156,9 +155,9 @@ class PhotosCollectionViewCell(itemView: View, private val emitter: PublishSubje
             )
         }
 
-        itemView.cardView_image_row_collection.post {
-            val w = itemView.cardView_image_row_collection.width
-            itemView.cardView_image_row_collection.apply {
+        binding.cardViewImageRowCollection.post {
+            val w = binding.cardViewImageRowCollection.width
+            binding.cardViewImageRowCollection.apply {
                 radius = (w/2).toDp(itemView.context).toFloat()
             }
         }
